feat: Rewrote radio buttons to the new renderer

+++ src/Renderer/Messaging/RenderMessage.h
@@ -14,7 +14,7 @@ class RenderMessage
 private:
     std::any data;
     Commands messageId = Commands::None;
-    long long int receiverId = -1;
+    size_t receiverId = -1;
     SubCommands subMessageId = SubCommands::None;
     RenderMessageSender *sender = nullptr;
 public:
@@ -24,12 +24,12 @@ public:
         return sender;
     }
 
-    [[nodiscard]] const long long int & GetReceiverId() const
+    [[nodiscard]] const size_t & GetReceiverId() const
     {
         return receiverId;
     }
 
-    void SetReceiverId(long long int receiverId)
+    void SetReceiverId(size_t receiverId)
     {
         RenderMessage::receiverId = receiverId;
     }
